(Test + Networking + Helpers) move mocks and helpers to own folder
