+++ test/helpers/mocks/mock_exception.dart
@@ -1,0 +1,9 @@
+/// Mock exception used for testing
+class MockException implements Exception {
+  final String message;
+
+  const MockException(this.message);
+
+  @override
+  String toString() => message;
+}
